requirements.txt
Flask==3.0.3
opencv-python-headless==4.8.1.78
numpy==1.26.4
Pillow==10.0.1
gunicorn==21.2.0

